Added preact support
