+++ overstate.d.ts
@@ -1,2 +1,10 @@
 // https://github.com/Microsoft/TypeScript/issues/2709#issuecomment-230182278
 declare module 'rollup-plugin-*';
+
+import { Store } from './store';
+
+export interface AppOptions<T extends {}> {
+  store: Store<T>;
+  view(props: { model: T }): any;
+  throttle?(fn: Function): any;
+}
