+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.1.0",
+  "version": "1.2.0",
   "name": "overstate",
   "description": "Infinitely composable state + actions 🎈",
   "keywords": [
